Budgets

+++ src/main/kotlin/com/kotlin/expensetracker/models/Budget.kt
@@ -2,4 +2,4 @@ package com.kotlin.expensetracker.models
 
 import java.util.Date
 
-data class Budget(val id: String?, val name: String, val createdAt: Date, val amount: Float)
+data class Budget(val id: String?, val name: String, val createdAt: Date, val amount: Double)
